Move bus tests into a dedicated unit
