+++ tests/InteractsWithMessengerTest.php
@@ -70,61 +70,6 @@ final class InteractsWithMessengerTest extends WebTestCase
     /**
      * @test
      */
-    public function can_interact_with_buses()
-    {
-        self::bootKernel();
-
-        self::getContainer()->get(MessageBusInterface::class)->dispatch(new MessageA());
-        $this->bus()->dispatched()
-            ->assertCount(1)
-            ->assertContains(MessageA::class, 1);
-    }
-
-    /**
-     * @test
-     */
-    public function use_default_sync_transport()
-    {
-        self::bootKernel(['environment' => 'default_sync_transport']);
-
-        self::getContainer()->get(MessageBusInterface::class)->dispatch(new MessageA());
-        $this->bus()->dispatched()
-            ->assertCount(1)
-            ->assertContains(MessageA::class, 1);
-    }
-
-    /**
-     * @test
-     */
-    public function interacts_with_specified_bus(): void
-    {
-        self::bootKernel(['environment' => 'multi_bus']);
-
-        $this->bus('bus_a')->dispatched()->assertEmpty();
-        $this->bus('bus_b')->dispatched()->assertEmpty();
-        $this->bus('bus_c')->dispatched()->assertEmpty();
-
-        self::getContainer()->get('bus_a')->dispatch(new MessageA(fail: true));
-        self::getContainer()->get('bus_b')->dispatch(new MessageB());
-        self::getContainer()->get('bus_c')->dispatch(new MessageC());
-
-        $this->transport()
-            ->process()
-            ->rejected()
-            ->assertContains(MessageA::class, 4)
-        ;
-
-        $this->bus('bus_a')->dispatched()->assertCount(1);
-        $this->bus('bus_b')->dispatched()->assertCount(1);
-        $this->bus('bus_c')->dispatched()->assertCount(1);
-        $this->bus('bus_a')->dispatched()->assertContains(MessageA::class, 1);
-        $this->bus('bus_b')->dispatched()->assertContains(MessageB::class, 1);
-        $this->bus('bus_c')->dispatched()->assertContains(MessageC::class, 1);
-    }
-
-    /**
-     * @test
-     */
     public function can_interact_with_queue(): void
     {
         self::bootKernel();
